pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "cpi"
version = "0.1.0"
description = "Confident Partial Interpretability — measurement scaffolding (pre-empirical)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["metrics", "experiments"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
